LayoutTests/http/tests/security/mixedcontent-geolocation-block-insecure-content-expected.txt

CONSOLE MESSAGE: line 14: [blocked] The page at https://127.0.0.1:8443/security/resources/geolocation-over-mixed-content-block.html was not allowed to display insecure content from http://127.0.0.1:8080/security/resources/compass.jpg.

This test loads a secure frame with insecure content that tries to access geolocation before loading insecure content. Access to insecure content is blocked over secure connections when geolocation is accessed.
